# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.EmployeeAttributes:
    properties:
      card_number_id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      warehouse_id:
        type: integer
    type: object
  domain.EmployeeDoc:
    properties:
      card_number_id:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      warehouse_id:
        type: integer
    type: object
  domain.EmployeeResponse:
    properties:
      data:
        $ref: '#/definitions/domain.EmployeeDoc'
    type: object
  domain.EmployeeWithInboundOrdersCount:
    properties:
      card_number_id:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      inbound_orders_count:
        type: integer
      last_name:
        type: string
      warehouse_id:
        type: string
    type: object
  domain.EmployeeWithInboundOrdersCountResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.EmployeeWithInboundOrdersCount'
        type: array
    type: object
  domain.EmployeesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.EmployeeDoc'
        type: array
    type: object
  domain.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestionar sellers y warehouses
  title: API de Sellers y Warehouses
  version: "1.0"
paths:
...
  /employee/reportInboundOrders:
    get:
      description: Reporte de inbound orders por empleado
      parameters:
      - description: ID de la sección
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reporte de inbound orders por empleado
          schema:
            $ref: '#/definitions/domain.EmployeeWithInboundOrdersCountResponse'
        "400":
          description: ID malformado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: El empleado no existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Reporte de inbound orders por empleado
      tags:
      - Employee
  /employees:
    get:
      description: Retorna una lista de todos los Employees registrados en el sistema
      produces:
      - application/json
      responses:
        "200":
          description: Listado de Employees
          schema:
            $ref: '#/definitions/domain.EmployeesResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Lista todos los Employees
      tags:
      - Employee
    post:
      consumes:
      - application/json
      description: Crea un nuevo empleado en el sistema con los datos proporcionados
      parameters:
      - description: Datos del empleado a crear
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/domain.EmployeeAttributes'
      produces:
      - application/json
      responses:
        "201":
          description: Empleado creado exitosamente
          schema:
            $ref: '#/definitions/domain.EmployeeResponse'
        "409":
          description: El código de tarjeta ya existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Datos malformados o faltantes
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Crea un nuevo empleado
      tags:
      - Employee
  /employees/{id}:
    delete:
      description: Elimina un employee por su ID de la base de datos
      parameters:
      - description: ID del employee
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: employee eliminado exitosamente
        "400":
          description: ID no enviado o malformado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: El employee no existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Elimina un employee por su ID
      tags:
      - Employee
    get:
      description: Retorna la información de un Employee por su ID
      parameters:
      - description: ID del Employee
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Listado de Employees
          schema:
            $ref: '#/definitions/domain.EmployeesResponse'
        "400":
          description: ID no enviado o malformado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: El employee no existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Retorna un Employee por su ID
      tags:
      - Employee
    patch:
      consumes:
      - application/json
      description: Actualiza los datos de un empleado existente en el sistema
      parameters:
      - description: ID del empleado
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del empleado a actualizar
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/domain.EmployeeAttributes'
      produces:
      - application/json
      responses:
        "200":
          description: Empleado actualizado exitosamente
          schema:
            $ref: '#/definitions/domain.EmployeeResponse'
        "400":
          description: ID no enviado o malformado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: El empleado no existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: El código de tarjeta ya existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Datos malformados o faltantes
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Actualiza un empleado por su ID
      tags:
      - Employee
...
swagger: "2.0"